requirements.txt
sensirion-i2c-driver
awscrt
awsiotsdk
getmac
PyYaml